Extracted dataset part from Glue
