+++ backend/dataall/aws/handlers/glue.py
@@ -5,8 +5,6 @@ from botocore.exceptions import ClientError
 from .service_handlers import Worker
 from .sts import SessionHelper
 from ...db import models
-from dataall.modules.datasets.db.models import DatasetTable, Dataset
-from dataall.modules.datasets.services.dataset_service import DatasetService
 
 log = logging.getLogger('aws:glue')
 
@@ -371,142 +369,6 @@ class Glue:
             raise e
 
     @staticmethod
-    @Worker.handler(path='glue.dataset.crawler.create')
-    def create_crawler(engine, task: models.Task):
-        with engine.scoped_session() as session:
-            dataset: Dataset = DatasetService.get_dataset_by_uri(
-                session, task.targetUri
-            )
-            location = task.payload.get('location')
-            Glue.create_glue_crawler(
-                **{
-                    'crawler_name': f'{dataset.GlueDatabaseName}-{location}'[:52],
-                    'region': dataset.region,
-                    'accountid': dataset.AwsAccountId,
-                    'database': dataset.GlueDatabaseName,
-                    'location': location or f's3://{dataset.S3BucketName}',
-                }
-            )
-
-    @staticmethod
-    def create_glue_crawler(**data):
-        try:
-            accountid = data['accountid']
-            database = data.get('database')
-            session = SessionHelper.remote_session(accountid=accountid)
-            glue = session.client('glue', region_name=data.get('region', 'eu-west-1'))
-            crawler_name = data.get('crawler_name')
-            targets = {'S3Targets': [{'Path': data.get('location')}]}
-            crawler = Glue._get_crawler(glue, crawler_name)
-            if crawler:
-                Glue._update_existing_crawler(
-                    glue, accountid, crawler_name, targets, database
-                )
-            else:
-                crawler = glue.create_crawler(
-                    Name=crawler_name,
-                    Role=SessionHelper.get_delegation_role_arn(accountid=accountid),
-                    DatabaseName=database,
-                    Targets=targets,
-                    Tags=data.get('tags', {'Application': 'dataall'}),
-                )
-
-            glue.start_crawler(Name=crawler_name)
-            log.info('Crawler %s started ', crawler_name)
-            return crawler
-        except ClientError as e:
-            log.error('Failed to create Crawler due to %s', e)
-
-    @staticmethod
-    def get_glue_crawler(data):
-        try:
-            accountid = data['accountid']
-            session = SessionHelper.remote_session(accountid=accountid)
-            glue = session.client('glue', region_name=data.get('region', 'eu-west-1'))
-            crawler_name = data.get('crawler_name')
-            crawler = Glue._get_crawler(glue, crawler_name)
-            return crawler
-        except ClientError as e:
-            log.error('Failed to find Crawler due to %s', e)
-            raise e
-
-    @staticmethod
-    @Worker.handler(path='glue.crawler.start')
-    def start_crawler(engine, task: models.Task):
-        with engine.scoped_session() as session:
-            dataset: Dataset = DatasetService.get_dataset_by_uri(
-                session, task.targetUri
-            )
-            location = task.payload.get('location')
-            return Glue.start_glue_crawler(
-                {
-                    'crawler_name': dataset.GlueCrawlerName,
-                    'region': dataset.region,
-                    'accountid': dataset.AwsAccountId,
-                    'database': dataset.GlueDatabaseName,
-                    'location': location,
-                }
-            )
-
-    @staticmethod
-    def start_glue_crawler(data):
-        try:
-            accountid = data['accountid']
-            crawler_name = data['crawler_name']
-            database = data['database']
-            targets = {'S3Targets': [{'Path': data.get('location')}]}
-            session = SessionHelper.remote_session(accountid=accountid)
-            glue = session.client('glue', region_name=data.get('region', 'eu-west-1'))
-            if data.get('location'):
-                Glue._update_existing_crawler(
-                    glue, accountid, crawler_name, targets, database
-                )
-            crawler = Glue._get_crawler(glue, crawler_name)
-            glue.start_crawler(Name=crawler_name)
-            log.info('Crawler %s started ', crawler_name)
-            return crawler
-        except ClientError as e:
-            log.error('Failed to start Crawler due to %s', e)
-            raise e
-
-    @staticmethod
-    def _get_crawler(glue, crawler_name):
-        crawler = None
-        try:
-            crawler = glue.get_crawler(Name=crawler_name)
-        except ClientError as e:
-            if e.response['Error']['Code'] == 'EntityNotFoundException':
-                log.debug(f'Crawler does not exists {crawler_name} %s', e)
-            else:
-                raise e
-        return crawler.get('Crawler') if crawler else None
-
-    @staticmethod
-    def _update_existing_crawler(glue, accountid, crawler_name, targets, database):
-        try:
-            glue.stop_crawler(Name=crawler_name)
-        except ClientError as e:
-            if (
-                e.response['Error']['Code'] == 'CrawlerStoppingException'
-                or e.response['Error']['Code'] == 'CrawlerNotRunningException'
-            ):
-                log.error('Failed to stop crawler %s', e)
-        try:
-            glue.update_crawler(
-                Name=crawler_name,
-                Role=SessionHelper.get_delegation_role_arn(accountid=accountid),
-                DatabaseName=database,
-                Targets=targets,
-            )
-            log.info('Crawler %s updated ', crawler_name)
-        except ClientError as e:
-            log.debug('Failed to stop and update crawler %s', e)
-            if e.response['Error']['Code'] != 'CrawlerRunningException':
-                log.error('Failed to update crawler %s', e)
-            else:
-                raise e
-
-    @staticmethod
     @Worker.handler(path='glue.job.runs')
     def get_job_runs(engine, task: models.Task):
         with engine.scoped_session() as session:
@@ -522,44 +384,3 @@ class Glue:
                 log.warning(f'Could not retrieve pipeline runs , {str(e)}')
                 return []
             return response['JobRuns']
-
-    @staticmethod
-    def grant_principals_all_table_permissions(
-        table: DatasetTable, principals: [str], client=None
-    ):
-        """
-        Update the table permissions on Lake Formation
-        for tables managed by data.all
-        :param principals:
-        :param table:
-        :param client:
-        :return:
-        """
-        if not client:
-            client = SessionHelper.remote_session(table.AWSAccountId).client(
-                'lakeformation', region_name=table.region
-            )
-        for principal in principals:
-            try:
-                grant_dict = dict(
-                    Principal={'DataLakePrincipalIdentifier': principal},
-                    Resource={
-                        'Table': {
-                            'DatabaseName': table.GlueDatabaseName,
-                            'Name': table.name,
-                        }
-                    },
-                    Permissions=['ALL'],
-                )
-                response = client.grant_permissions(**grant_dict)
-                log.error(
-                    f'Successfully granted principals {principals} all permissions on table '
-                    f'aws://{table.AWSAccountId}/{table.GlueDatabaseName}/{table.name} '
-                    f'access: {response}'
-                )
-            except ClientError as e:
-                log.error(
-                    f'Failed to grant admin roles {principals} all permissions on table '
-                    f'aws://{table.AWSAccountId}/{table.GlueDatabaseName}/{table.name} '
-                    f'access: {e}'
-                )
